Adding a test on empty list before removing files in codegen.
git-svn-id: http://svn.code.sf.net/p/preesm/code/trunk@1659 c7aee7ca-e32f-b942-b52e-c1c222c4573e

+++ plugins/org.ietr.preesm.codegen.xtend/src/org/ietr/preesm/codegen/xtend/task/CodegenTask.java
@@ -190,12 +190,14 @@ public class CodegenTask extends AbstractTaskImplementation {
 			IFolder f = workspace.getRoot().getFolder(new Path(codegenPath));
 			File folder = new File(f.getRawLocation().toOSString());
 			File fList[] = folder.listFiles();
-			// Searches .extension
-			for (int i = 0; i < fList.length; i++) {
-				String pes = fList[i].getName();
-				if (pes.endsWith(extension)) {
-					// and deletes
-					fList[i].delete();
+			if(fList != null){
+				// Searches .extension
+				for (int i = 0; i < fList.length; i++) {
+					String pes = fList[i].getName();
+					if (pes.endsWith(extension)) {
+						// and deletes
+						fList[i].delete();
+					}
 				}
 			}
